feat: 상품 상세 조회 반환값에 옵션,카테고리,리뷰 추가#33

+++ src/main/java/spharos/msg/domain/product/dto/ProductResponse.java
@@ -41,13 +41,13 @@ public class ProductResponse {
         @Schema(description = "상품 이미지")
         private String image;
         @Schema(description = "상품 별점")
-        private Short productStar;
+        private BigDecimal productStar;
         @Schema(description = "상품 할인율")
         private BigDecimal discountRate;
 
         @Builder
         private ProductInfo(Long productId, String productName, Integer productPrice, String image,
-            Short productStar, BigDecimal discountRate) {
+            BigDecimal productStar, BigDecimal discountRate) {
 
             this.productId = productId;
             this.productName = productName;
